Fix overflow: use ll for edge weights and total cost

+++ content/graphs/kruskal.cpp
@@ -3,14 +3,14 @@
  *Status:* Tested
 */
 struct Edge {
-  int a; int b; int w;
-  Edge(int a_, int b_, int w_) : a(a_), b(b_), w(w_) {}
+  int a; int b; ll w;
+  Edge(int a_, int b_, ll w_) : a(a_), b(b_), w(w_) {}
 };
 bool c_edge(Edge &a, Edge &b) { return a.w < b.w; }
-int Kruskal() {
+ll Kruskal() {
   int n = G.size();
   union_find sets(n);
-  vector< Edge > edges;
+  vector<Edge> edges;
   for(int i = 0; i < n; i++) {
     for(auto eg : G[i]) {
       // node i to node eg.first with cost eg.second
@@ -18,7 +18,7 @@ int Kruskal() {
     }
   }
   sort(edges.begin(), edges.end(), c_edge);
-  int min_cost = 0;
+  ll min_cost = 0;
   for(Edge e : edges) {
     if(sets.sameSet(e.a, e.b) != true) {
       tree.emplace_back(e.a, e.b, e.w);
